[phi] Decouple phi/core from fluid: move IPU device code into phi/core
Move IPU device query code from paddle/fluid/platform/device/ipu/ into
paddle/phi/core/platform/device/ipu/, consistent with how GPU and XPU
device code is organized. This eliminates phi/core's runtime dependency
on fluid for IPU device counting.

Changes:
- New: paddle/phi/core/platform/device/ipu/{ipu_device,ipu_info}.{h,cc}
- New: paddle/phi/core/platform/device_count.{h,cc} (generic device
  count provider registry for future extensibility)
- allocator_facade.cc now uses platform::GetIPUDeviceCount() directly
  (consistent with XPU pattern)
- Old fluid locations become forwarding headers for backward compat
- CMake: phi_core links popart-only when WITH_IPU; fluid ipu_info
  target becomes a thin compatibility shim depending on phi
