+++ paddle/fluid/platform/device/ipu/ipu_info.h
@@ -15,13 +15,6 @@ limitations under the License. */
 #include <vector>
 
 #include "glog/logging.h"
-
-namespace paddle {
-namespace platform {
-
-std::vector<int> GetSelectedIPUDevices();
-int GetIPUDeviceCount();
-
-}  // namespace platform
-}  // namespace paddle
 #endif
+
+#include "paddle/phi/core/platform/device/ipu/ipu_info.h"
