+++ paddle/phi/core/platform/device/ipu/ipu_device.h
@@ -1,0 +1,33 @@
+// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#ifdef PADDLE_WITH_IPU
+#include <vector>
+
+namespace paddle {
+namespace platform {
+namespace ipu {
+
+// Get the number of all available IPU devices.
+int GetNumDevices();
+
+// Get the device id of all available IPU devices.
+std::vector<int> GetDeviceIds();
+
+}  // namespace ipu
+}  // namespace platform
+}  // namespace paddle
+#endif
