+++ paddle/phi/core/platform/device/ipu/ipu_info.cc
@@ -1,0 +1,29 @@
+// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "paddle/phi/core/platform/device/ipu/ipu_info.h"
+
+#ifdef PADDLE_WITH_IPU
+#include "paddle/phi/core/platform/device/ipu/ipu_device.h"
+
+namespace paddle {
+namespace platform {
+
+std::vector<int> GetSelectedIPUDevices() { return ipu::GetDeviceIds(); }
+
+int GetIPUDeviceCount() { return ipu::GetNumDevices(); }
+
+}  // namespace platform
+}  // namespace paddle
+#endif
